requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.2
idna==2.10
insta-scrape==2.1.2
multidict==5.1.0
pycodestyle==2.7.0
python-dotenv==0.17.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
